........
  r670689 | adelmelle | 2008-06-23 19:01:25 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Fix XMLObj.addCharacters(): incorrect after r670341
........
  r670723 | adelmelle | 2008-06-23 20:56:51 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Javadoc modification + addition of clarifying comments
........
  r670760 | adelmelle | 2008-06-23 22:18:45 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Javadoc fixup
........
  r670761 | adelmelle | 2008-06-23 22:21:09 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Minor cleanup: remove unused parameter from method signature
........
  r670763 | adelmelle | 2008-06-23 22:22:48 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Minor cleanup: remove double storage of the FObj
........
  r670765 | adelmelle | 2008-06-23 22:24:46 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Minor cleanup: remove redundant cast
........
  r670766 | adelmelle | 2008-06-23 22:25:30 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Minor cleanup: remove redundant cast
........
  r670767 | adelmelle | 2008-06-23 22:26:56 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Javadoc fixup
........
  r670770 | adelmelle | 2008-06-23 22:29:53 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Minor cleanup: simplified conditional
........
  r670773 | adelmelle | 2008-06-23 22:32:31 +0100 (Mon, 23 Jun 2008) | 1 line
  
  A nit...
........
  r670777 | adelmelle | 2008-06-23 22:39:19 +0100 (Mon, 23 Jun 2008) | 1 line
  
  Fix imports...
........
  r671145 | maxberger | 2008-06-24 12:59:33 +0100 (Tue, 24 Jun 2008) | 2 lines
  
  Another character handling problem, replaced end-start by length
........
  r672010 | adelmelle | 2008-06-26 20:44:20 +0100 (Thu, 26 Jun 2008) | 1 line
  
  Fixed a possible NullPointerException: if the citation appears in a fo:marker, the property will only be parsed if the marker is retrieved...
........
  r672026 | acumiskey | 2008-06-26 21:49:11 +0100 (Thu, 26 Jun 2008) | 5 lines
  
  Some basic cleanup.
  
  * Duplicated/redundant PSDictionary, PSPageDeviceDictionary and PSDictionaryFormatException removed and now referenced from xmlgraphics commons.
  * Updated xmlgraphics commons jar containing migrated PSPageDeviceDictionary class.
........


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/branches/Temp_AFPGOCAResources@672235 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/layoutmgr/FlowLayoutManager.java
@@ -68,7 +68,7 @@ public class FlowLayoutManager extends BlockStackingLayoutManager
 
         // set layout dimensions
         int flowIPD = getCurrentPV().getCurrentSpan().getColumnWidth();
-        int flowBPD = (int) getCurrentPV().getBodyRegion().getBPD();
+        int flowBPD = getCurrentPV().getBodyRegion().getBPD();
 
         // currently active LM
         LayoutManager curLM;
